Updated the layout and fetch route from hook
